+++ HISTORY.rst
@@ -1,6 +1,10 @@
 Release History
 ===============
 
+Unreleased Changes
+------------------
+* Added a `download_and_unzip` function to the API.
+
 0.3.3 (2019/11/09)
 ------------------
 * Fixed an issue that would cause the download rate to be under estimated.